% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BobTestPackage.R
\docType{package}
\name{BobTestPackage-package}
\alias{BobTestPackage}
\alias{BobTestPackage-package}
\title{BobTestPackage: quick test of setup}
\description{
With Sequoia testing package setup.
}
\author{
\strong{Maintainer}: Bob Verity \email{rverity@imperial.ac.uk}

}
\keyword{internal}
